merge dev to test

See merge request tio/3n1-android!314

+++ app/src/androidTest/java/com/walisport/app/ExampleInstrumentedTest.kt
@@ -46,7 +46,7 @@ class ExampleInstrumentedTest {
                         Style.IconBadge(
                             R.mipmap.ic_fifa.getDrawable(),
                             "世界杯",
-                            (-18f).dp2px
+                            (-18f).dp2px.toInt()
                         )
                     )
                 } else {
